+++ src/test/java/com/alechilles/alecstamework/persistence/architecture/ReplacementPersistenceArchitectureGuardTest.java
@@ -324,6 +324,7 @@ class ReplacementPersistenceArchitectureGuardTest {
                         "activeCaptureFailureCooldown",
                         "diagnoseCoopCapture",
                         "diagnoseCoopRelease",
+                        "findAlias",
                         "findAllCommandRosters",
                         "findAllLifecycles",
                         "findAllPopulationGroupAssignments",